requirements.txt
PyYAML==6.0
Flask==2.2.2
